+++ R/has_color.r
@@ -5,26 +5,41 @@
 #'
 #' @details
 #' The following algorithm is used to detect ANSI support: \itemize{
-#'   \item If the \code{crayon.enabled} option is set to \code{TRUE}
-#'     with \code{options()}, then \code{TRUE} is returned. If it is
-#'     set to something else than \code{TRUE} (typically \code{FALSE}),
-#'     then \code{FALSE} is returned.
+#'   \item If the `crayon.enabled` option is set to `TRUE`
+#'     with [options()], then `TRUE` is returned. If it is
+#'     set to something else than `TRUE` (typically `FALSE`),
+#'     then `FALSE` is returned.
+#'   \item Otherwise, if the `NO_COLOR` environment variable is
+#'     set, `FALSE` is returned.
 #'   \item Otherwise, if the standard output is not a terminal, then
-#'     \code{FALSE} is returned.
-#'   \item Otherwise, if the platform is Windows, \code{TRUE} is returned
+#'     `FALSE` is returned.
+#'   \item Otherwise, if the platform is Windows, `TRUE` is returned
 #'     if running in ConEmu (\url{https://conemu.github.io/}) or
 #'     cmder (\url{http://cmder.net}) with ANSI color support.
-#'     Otherwise \code{FALSE} is returned.
-#'   \item Otherwise, if the \code{COLORTERM} environment variable is
-#'     set, \code{TRUE} is returned.
-#'   \item Otherwise, if the \code{TERM} environment variable starts
-#'     with \code{screen}, \code{xterm} or \code{vt100}, or matches
-#'     \code{color}, \code{ansi}, \code{cygwin} or \code{linux}
-#'     (with case insentive matching), then \code{TRUE} is returned.
-#'   \item Otherwise \code{FALSE} is returned.
+#'     Otherwise `FALSE` is returned.
+#'   \item Otherwise, if the `COLORTERM` environment variable is
+#'     set, `TRUE` is returned.
+#'   \item Otherwise, if the `TERM` environment variable starts
+#'     with `screen`, `xterm` or `vt100`, or matches
+#'     `color`, `ansi`, `cygwin` or `linux`
+#'     (with case insentive matching), then `TRUE` is returned.
+#'   \item Otherwise `FALSE` is returned.
 #' }
 #'
-#' @return \code{TRUE} if the current R session supports color.
+#' @section Sinks:
+#' Note that `has_color()` returns `FALSE` if a sink is active
+#' (see [sink()]). It assumes that the constructed string will be printed
+#' to the standard output, and `sink()` redirects to a file, and usually
+#' you don't want ANSI colors in the file.
+#'
+#' The same applies to the case when R's standard output is redirected
+#' to a file, from the command line, e.g.:
+#' ```
+#' R -q -e 'cat(crayon::red("no color here\n"))' > /tmp/crayon-test.txt
+#' cat /tmp/crayon-test.txt
+#' ```
+#'
+#' @return `TRUE` if the current R session supports color.
 #'
 #' @export
 #' @examples
@@ -36,20 +51,30 @@ has_color <- function() {
   enabled <- getOption("crayon.enabled")
   if (!is.null(enabled)) { return(isTRUE(enabled))  }
 
+  if (!is.na(Sys.getenv("NO_COLOR", NA))) {
+    return(FALSE)
+  }
+  
+  ## RStudio with (potential) ANSI support?
+  if (rstudio_with_ansi_support() && sink.number() == 0) {
+    return(TRUE)
+  }
+
   ## Are we in a terminal? No?
   if (!isatty(stdout())) { return(FALSE) }
 
   ## Are we in a windows terminal with color support?
-  if (.Platform$OS.type == "windows") {
+  if (os_type() == "windows") {
     if (Sys.getenv("ConEmuANSI") == "ON") { return(TRUE) }
     if (Sys.getenv("CMDER_ROOT") != "") { return(TRUE) }
+    if (Sys.getenv("ANSICON") != "") {  return(TRUE) }
 
     ## Are we in another windows terminal or GUI? :(
     return(FALSE)
   }
 
   ## Running in a recent Emacs?
-  if (inside_emacs() && emacs_version()[1] >= 23) { return(TRUE) }
+  if (inside_emacs() && ! is.na(emacs_version()[1]) && emacs_version()[1] >= 23) { return(TRUE) }
 
   ## COLORTERM set?
   if ("COLORTERM" %in% names(Sys.getenv())) { return(TRUE) }
@@ -65,26 +90,27 @@ has_color <- function() {
 #' Number of colors the terminal supports
 #'
 #' @details
-#' If the \code{crayon.colors} option is set, then we
+#' If the `crayon.colors` option is set, then we
 #' just use that. It should be an integer number. You can use this
 #' option as a workaround if crayon does not detect the number of
 #' colors accurately.
 #'
 #' In Emacs, we report eight colors.
 #'
-#' Otherwise, we use the \code{tput} shell command to detect the
-#' number of colors. If \code{tput} is not available,
+#' Otherwise, we use the `tput` shell command to detect the
+#' number of colors. If `tput` is not available,
 #' but we think that the terminal supports colors, then
 #' eigth colors are assumed.
 #'
 #' If tput returns 8, but TERM is xterm, we return 256, as most xterm
 #' compatible terminals in fact do support 256 colors.
 #' There is some discussion about this here:
-#' \url{https://github.com/gaborcsardi/crayon/issues/17}
+#' \url{https://github.com/r-lib/crayon/issues/17}
 #'
-#' For efficiency, \code{num_colors} caches its result. To
-#' re-check the number of colors, set the \code{forget} argument to
-#' \code{TRUE}.
+#' For efficiency, `num_colors()` caches its result. To
+#' re-check the number of colors, set the `forget` argument to
+#' `TRUE`. The cached value is only used if no sinks are active,
+#' see also [has_color()] for more information about sinks.
 #'
 #' @param forget Whether to forget the cached result of the color check.
 #' @return Numeric scalar, the number of colors the terminal supports.
@@ -98,8 +124,17 @@ num_colors <- (function() {
   cache <- NULL
 
   function(forget=FALSE) {
-    if (forget || is.null(cache)) cache <<- i_num_colors()
-    cache
+    if (forget || is.null(cache)) {
+      result <- i_num_colors()
+      if (sink.number() == 0 && rstudio_initialized()) {
+        cache <<- result
+      } else {
+        cache <<- NULL
+      }
+    } else {
+      result <- cache
+    }
+    result
   }
 })()
 
@@ -109,16 +144,37 @@ i_num_colors <- function() {
   cols <- getOption("crayon.colors")
   if (!is.null(cols)) { return(as.integer(cols)) }
 
-  ## Otherwise try to detect. If no color support, then 1
+  ## RStudio
+  if (rstudio_with_ansi_support()) { return(256) }
+
   if (!has_color()) { return(1) }
 
   ## Emacs
   if (inside_emacs()) { return(8) }
 
+  ## Are we in a windows terminal with color support?
+  if (os_type() == "windows") {
+    if (Sys.getenv("ConEmuANSI") == "ON" ||
+        Sys.getenv("CMDER_ROOT") != "") {
+      return(get_terminal_colors())
+    }
+    if (Sys.getenv("ANSICON") != "") return(8)
+
+    ## Are we in another windows terminal or GUI? :(
+    return(1)
+  }
+
+  ## Otherwise
+  get_terminal_colors()
+}
+
+get_terminal_colors <- function() {
   ## Try to run tput colors. If it did not run, but has_colors() is TRUE,
   ## then we just report 8 colors
-  cols <- suppressWarnings(try(silent = TRUE,
-              as.numeric(system("tput colors", intern = TRUE))))
+  cols <- suppressWarnings(try(
+    silent = TRUE,
+    as.numeric(system("tput colors 2>/dev/null", intern = TRUE))[1]
+  ))
   if (inherits(cols, "try-error") || !length(cols) || is.na(cols)) { return(8) }
   if (cols %in% c(-1, 0, 1)) { return(1) }