+++ R/string.r
@@ -1,11 +1,11 @@
 
 #' Convert to character
 #'
-#' This function just calls \code{as.character}, but it is
+#' This function just calls [as.character()], but it is
 #' easier to type and read.
 #'
 #' @param x Object to be coerced.
-#' @param ... Further arguments to pass to \code{as.character}.
+#' @param ... Further arguments to pass to `as.character()`.
 #' @return Character value.
 #'
 #' @export
@@ -54,4 +54,24 @@ chr <- function(x, ...) as.character(x, ...)
   } else {
     paste0(lhs, rhs)
   }
+}
+
+#' @export
+as.character.crayon <- function(x, ...) {
+  start(x)
+}
+
+#' @export
+`-.crayon` <- function(e1, e2) {
+  if (!missing(e2)) {
+    base::`-`(e1, e2)
+  }
+  my_styles <- attr(e1, "_styles")
+  if (has_color()) {
+    for (i in seq_along(my_styles)) {
+      my_styles[[i]]$open <- my_styles[[i]]$close
+    }
+  }
+  attr(e1, "_styles") <- my_styles
+  e1
 }